selfattest_web/requirements-dev.txt
Django==3.0.5
djangorestframework==3.11.0
python-decouple==3.3
psycopg2-binary==2.7.5
dj-database-url==0.5.0
django-cors-headers==3.2.1

coverage==5.0.4 #Test

gunicorn==20.0.4 #App server

redis==3.5.2 #cache/ratelimit

mypy==0.770 #Dev only, for type checking.
django-debug-toolbar==2.2


sentry-sdk==0.14.4 #Error tracker

django-storages==1.9.1 #DO space storage 
boto3==1.14.1 #S3 connect

Pillow==7.1.2 #Images

django-redis==4.12.1 #Caching. Required for Rate limit at least.

django-silk==4.0.1 #https://github.com/jazzband/django-silk. Live Profiling

